project/app/src/main/java/com/nexte/nexte/ObjectModels/Comment.kt
package com.nexte.nexte.ObjectModels

import java.util.*

class Comment(val id: String,
              val userId: String,
              val comment: String,
              val date: Date) {

    enum class ServerRequest(val request: Map<String, String>) {
        COMMENTS(hashMapOf("route" to "comments", "method" to "get")),
        ADD_COMMENT(hashMapOf("route" to "comment", "method" to "post")),
        DELETE_COMMENT(hashMapOf("route" to "comment", "method" to "delete")),
        REPORT_COMMENT(hashMapOf("route" to "report-comment", "method" to "post"))
    }
}